Finish 2 colum header and footer layout
It's made with a wrapper which for some reason is
necessary for the webpage not to rip and show the
bare raw nothing.
flex-direction: row also creates gaps, and for a
while vertical-center didn't strech to fit the
body. Then it fixed itself by...  using the
wrapper, i guess? I don't understand css i'm
failing this class dude wth